feat(dashboard): refactor guest and tenant sidebars; implement new sidebar component structure and navigation links
